requirements.txt
Pillow==5.00
django==2.0.2
django-ckeditor==5.4.0